+++ crates/sargon/src/profile/v100/mod.rs
@@ -7,6 +7,7 @@ mod header;
 mod networks;
 mod profile;
 mod profile_file_contents;
+mod profile_legacy_state_bugs;
 mod proto_profile_maybe_with_legacy_p2p_links;
 
 pub use address::*;
@@ -18,4 +19,5 @@ pub use header::*;
 pub use networks::*;
 pub use profile::*;
 pub use profile_file_contents::*;
+pub use profile_legacy_state_bugs::*;
 pub use proto_profile_maybe_with_legacy_p2p_links::*;
